+++ src/compile/Compiler.ts
@@ -117,6 +117,8 @@ export default class Compiler {
 	expectedProperties: Set<string>;
 	usesRefs: boolean;
 
+	file: string;
+	fileVar: string;
 	locate: (c: number) => { line: number, column: number };
 
 	stylesheet: Stylesheet;
@@ -159,6 +161,9 @@ export default class Compiler {
 		this.bindingGroups = [];
 		this.indirectDependencies = new Map();
 
+		this.file = options.filename && (
+			typeof process !== 'undefined' ? options.filename.replace(process.cwd(), '').replace(/^[\/\\]/, '') : options.filename
+		);
 		this.locate = getLocator(this.source);
 
 		// track which properties are needed, so we can provide useful info
@@ -177,6 +182,8 @@ export default class Compiler {
 		this.templateVars = new Map();
 		this.aliases = new Map();
 		this.usedNames = new Set();
+
+		this.fileVar = options.dev && this.getUniqueName('file');
 
 		this.computations = [];
 		this.templateProperties = {};