+++ src/shared/index.js
@@ -65,7 +65,7 @@ export function init(component, options) {
 
 	component.options = options;
 	component.root = options.root || component;
-	component.store = component.root.store || options.store;
+	component.store = options.store || component.root.store;
 }
 
 export function on(eventName, handler) {