+++ src/shared/utils.js
@@ -15,7 +15,14 @@ export function isPromise(value) {
 }
 
 export function callAfter(fn, i) {
+	if (i === 0) fn();
 	return () => {
 		if (!--i) fn();
+	};
+}
+
+export function addLoc(element, file, line, column, char) {
+	element.__svelte_meta = {
+		loc: { file, line, column, char }
 	};
 }